Update project handle multiple extensions
I need another extension for my workflow with my Dave Smith Mopho.
I like working in mono repos wherever possible. I've broken out the
project into three sub projects. MPD and UM-One are the extensions, and
Common is for all the shared crap.

The extensions will build basically the same, so I moved the extension
specific stuff into base-extension.gradle. That way I wont have repeated
code in the sub projects gradle.
