+++ base-extension.gradle
@@ -1,0 +1,15 @@
+dependencies {
+    implementation project(":common")
+}
+
+jar {
+    from {
+        configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
+    }
+}
+
+task install(type: Copy) {
+    from jar
+    into "${System.env.BITWIG_EXTENSIONS_LOCATION}"
+    rename {"${project.name}.bwextension"}
+}
