+++ build.gradle
@@ -1,43 +1,34 @@
+allprojects {
+    repositories {
+        mavenCentral()
+        jcenter()
+        maven {
+            url 'https://maven.bitwig.com'
+        }
+    }
+}
+
+subprojects {
+    apply plugin: "java"
+    apply plugin: "org.jetbrains.kotlin.jvm"
+
+    compileJava {
+        sourceCompatibility = 1.8
+        targetCompatibility = 1.8
+    }
+}
+
 buildscript {
+    repositories {
+        mavenCentral()
+        jcenter()
+        maven {
+            url 'https://maven.bitwig.com'
+        }
+    }
+
     dependencies {
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.50"
     }
-}
 
-plugins {
-    id 'java'
-    id 'org.jetbrains.kotlin.jvm' version "1.3.50"
-}
-
-repositories {
-    mavenCentral()
-    jcenter()
-    maven {
-        url 'https://maven.bitwig.com'
-    }
-}
-
-dependencies {
-    compile 'org.jetbrains.kotlin:kotlin-stdlib'
-    compile 'org.jetbrains.kotlin:kotlin-stdlib-jdk8'
-    compile 'org.jetbrains.kotlin:kotlin-reflect'
-    compile 'com.bitwig:extension-api:9+'
-    compile 'com.google.inject:guice:4.1.0'
-}
-
-compileJava {
-    sourceCompatibility = 1.8
-    targetCompatibility = 1.8
-}
-
-jar {
-    from {
-        configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
-    }
-}
-
-task install(type: Copy) {
-    from jar
-    into "${System.env.BITWIG_EXTENSIONS_LOCATION}"
-    rename {"mpd.bwextension"}
 }
