+++ common/src/main/kotlin/offthecob/common/CommonExtensionDefinition.kt
@@ -1,0 +1,4 @@
+package offthecob.common
+
+class CommonExtensionDefinition {
+}
