+++ mpd/mpd.gradle
@@ -1,0 +1,3 @@
+apply from: "../base-extension.gradle"
+
+
