+++ mpd/src/main/kotlin/com/offthecob/Mode.kt
