+++ mpd/src/main/kotlin/com/offthecob/MpdModule.kt
