+++ mpd/src/main/resources/META-INF/services/com.bitwig.extension.ExtensionDefinition
