+++ settings.gradle
@@ -1,0 +1,6 @@
+include 'common', 'mpd', 'um-one'
+
+rootProject.children.each { project ->
+    project.buildFileName = "${project.name}.gradle"
+    assert project.buildFile.isFile()
+}
