--- src/main/kotlin/com/offthecob/Navigation.kt
@@ -1,15 +1,0 @@
-package com.offthecob
-
-import com.bitwig.extension.controller.api.Arranger
-import com.google.inject.Inject
-
-class Navigation @Inject constructor(private val arranger: Arranger) {
-
-    init {
-        arranger.isClipLauncherVisible.markInterested()
-    }
-
-    fun toggleClipLauncher() {
-        arranger.isClipLauncherVisible.toggle()
-    }
-}
