--- src/main/kotlin/com/offthecob/constants.kt
@@ -1,28 +1,0 @@
-package com.offthecob
-
-const val MPD_BAND_A: String = "f04700684000020601f7"
-const val MPD_BANK_B: String = "f04700684000020701f7"
-const val MPD_BANK_C: String = "f04700684000020801f7"
-const val REWIND: String = "f07f7f0605f7"
-const val FAST_FORWARD: String = "f07f7f0604f7"
-const val STOP: String = "f07f7f0601f7"
-const val PLAY: String = "f07f7f0602f7"
-const val RECORD: String = "f07f7f0606f7"
-const val PREVIEW_PRESS: String = "f04700684000021001f7"
-const val PREVIEW_RELEASE: String = "f04700684000021000f7"
-
-const val SLIDER_1: Int = 16
-const val SLIDER_2: Int = 17
-const val SLIDER_3: Int = 18
-const val SLIDER_4: Int = 19
-const val SLIDER_5: Int = 20
-const val SLIDER_6: Int = 21
-
-const val KNOB_1: Int = 30
-const val KNOB_2: Int = 31
-const val KNOB_3: Int = 32
-const val KNOB_4: Int = 33
-const val KNOB_5: Int = 34
-const val KNOB_6: Int = 35
-const val KNOB_7: Int = 36
-const val KNOB_8: Int = 37
