+++ um-one/src/main/resources/META-INF/services/com.bitwig.extension.ExtensionDefinition
@@ -1,0 +1,1 @@
+offthecob.um.one.UmOneExtensionDefinition
