+++ um-one/um-one.gradle
@@ -1,0 +1,1 @@
+apply from: "../base-extension.gradle"
